requirements.txt
psycopg2==2.9.3
Django==4.0.4
Pillow==9.1.0
django-durationwidget==1.0.5